% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a11y.R
\name{a11y_css}
\alias{a11y_css}
\alias{html_dependency_a11y_css}
\title{Test your document or app for accessibility issues.}
\usage{
a11y_css(
  language = c("en", "ar", "fr", "gr", "pt-br", "ru", "zh"),
  level = c("all", "error", "obsolete", "warning")
)

html_dependency_a11y_css(
  language = c("en", "ar", "fr", "gr", "pt-br", "ru", "zh"),
  level = c("all", "error", "obsolete", "warning")
)
}
\arguments{
\item{language}{Your desired language}

\item{level}{The level of errors:
\itemize{
\item \code{all}: every messages (called \code{advice} by \code{a11y.css})
\item \code{warning}: warnings and errors;
\item \code{obsolete}: obsolete stuff, warnings and errors;
\item \code{error}: only errors.
}}
}
\value{
An \code{htmltools} tag list with the \code{a11y.css} dependency.
}
\description{
Uses \href{https://github.com/ffoodd/a11y.css}{a11y.css} to include warnings and
diagnostics about potential accessibility issues in your HTML document or
Shiny web app.
}
\section{Functions}{
\itemize{
\item \code{html_dependency_a11y_css}: The \code{a11y.css} html dependency alone
}}

